Add files via upload
The task is to build a system that lets users ask questions about information from websites. The system will gather data from websites, break it into smaller pieces, and store it in a way that makes it easy to search. When a user asks a question, the system will find the most relevant information and use a language model to give a clear and accurate answer based on the data from the websites.